+++ workspaces/redhat-argocd/plugins/argocd/src/components/DeploymentLifeCycle/sidebar/resources/ResourcesTable.tsx
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2024 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import React, { FC, useCallback, useState } from 'react';
 import {
   makeStyles,
@@ -11,9 +26,14 @@ import {
 import { ResourcesTableBody } from './ResourcesTableBody';
 import { ResourcesTableHeader } from './ResourcesTableHeader';
 import { ResourcesColumnHeaders } from './ResourcesColumnHeader';
-import { ResourcesSearchBar } from './ResourcesSearchBar';
-import { ResourcesFilterBy } from './ResourcesFilterBy';
+import { ResourcesFilterBy } from './filters/ResourcesFilterBy';
 import { Order, Resource } from '../../../../types/application';
+import { FiltersType } from '../../../../types/resources';
+import {
+  getResourceCreateTimestamp,
+  sortValues,
+} from '../../../../utils/utils';
+import { useArgoResources } from '../rollouts/RolloutContext';
 
 interface ResourcesTableProps {
   resources: Resource[];
@@ -24,6 +44,7 @@ const useStyles = makeStyles(theme => ({
   table: {
     width: '100%',
     marginTop: theme.spacing(2),
+    minHeight: theme.spacing(20),
   },
   empty: {
     padding: theme.spacing(2),
@@ -43,78 +64,99 @@ export const ResourcesTable: FC<ResourcesTableProps> = ({
 }) => {
   const classes = useStyles();
 
-  const [searchValue, setSearchValue] = useState<string>('');
-  const [filterValue, setFilterValue] = useState<string>('');
   const [order, setOrder] = useState<Order>('desc');
   const [orderBy, setOrderBy] = useState<string | null>(null);
   const [orderById, setOrderById] = useState<string | null>(null);
   const [page, setPage] = useState<number>(0);
   const [rowsPerPage, setRowsPerPage] = useState<number>(5);
+  const [filters, setFilters] = useState<FiltersType>({
+    Kind: [],
+    HealthStatus: [],
+    SyncStatus: [],
+    SearchByName: [],
+  });
 
-  const getSortableRowValues = useCallback(
-    (res: Resource) => {
-      const {
-        kind,
-        status,
-        health: { status: healthStatus },
-      } = res;
+  const getSortableRowValues = useCallback((res: Resource) => {
+    const {
+      name = '',
+      kind = '',
+      status = '',
+      health = { status: '' },
+      createTimestamp,
+    } = res;
+    const healthStatus = health.status || '';
 
-      return [kind, createdAt, status, healthStatus];
-    },
-    [createdAt],
-  );
+    return [undefined, name, kind, createTimestamp, status, healthStatus];
+  }, []);
 
-  const allResources = React.useMemo(
-    () => resources.map((r, idx) => ({ ...r, id: idx })) ?? [],
-    [resources],
-  );
+  const { argoResources } = useArgoResources();
+
+  const allResources = React.useMemo(() => {
+    const getTimestamp = (row: Resource) =>
+      getResourceCreateTimestamp(argoResources, row) || createdAt;
+
+    return (
+      resources.map((r, idx) => ({
+        ...r,
+        id: idx,
+        createTimestamp: getTimestamp(r),
+      })) ?? []
+    );
+  }, [resources, argoResources, createdAt]);
+
+  const getAllKinds = React.useMemo(() => {
+    return resources.reduce((kinds: string[], resource) => {
+      if (resource.kind && !kinds.includes(resource.kind)) {
+        kinds.push(resource.kind);
+      }
+      return kinds;
+    }, []);
+  }, [resources]);
 
   const filteredResources = React.useMemo(() => {
     let items = allResources;
 
-    // Filter by health status
-    if (filterValue && filterValue !== 'All') {
-      items = items.filter(resource => resource.health.status === filterValue);
-    }
-
-    // Search by kind
-    if (searchValue) {
+    // Filters
+    if (filters.HealthStatus.length) {
       items = items.filter(resource =>
-        resource.kind
+        filters.HealthStatus.includes(resource.health?.status ?? ''),
+      );
+    }
+    if (filters.SyncStatus.length) {
+      items = items.filter(resource =>
+        filters.SyncStatus.includes(resource.status),
+      );
+    }
+    if (filters.Kind.length) {
+      items = items.filter(resource => filters.Kind.includes(resource.kind));
+    }
+    if (filters.SearchByName.length) {
+      items = items.filter(resource =>
+        resource.name
           .toLocaleLowerCase()
-          .includes(searchValue.toLocaleLowerCase()),
+          .includes(filters.SearchByName[0].toLocaleLowerCase()),
       );
     }
 
-    // Sort
+    // Sorting
     if (orderById && order) {
       items = items.sort((a, b) => {
-        const aValue = getSortableRowValues(a)[parseInt(orderById, 10)];
-        const bValue = getSortableRowValues(b)[parseInt(orderById, 10)];
+        let aValue = getSortableRowValues(a)[parseInt(orderById, 10)];
+        let bValue = getSortableRowValues(b)[parseInt(orderById, 10)];
 
         if (aValue === undefined || bValue === undefined) return 0;
 
-        if (typeof aValue === 'number' && typeof bValue === 'number') {
-          return order === 'asc' ? aValue - bValue : bValue - aValue;
+        if (orderBy === 'created-at') {
+          aValue = new Date(aValue).toISOString();
+          bValue = new Date(bValue).toISOString();
         }
-        if (typeof aValue === 'string' && typeof bValue === 'string') {
-          return order === 'asc'
-            ? aValue.localeCompare(bValue)
-            : bValue.localeCompare(aValue);
-        }
-        return 0;
+
+        return sortValues(aValue, bValue, order);
       });
     }
 
     return items;
-  }, [
-    allResources,
-    filterValue,
-    searchValue,
-    orderById,
-    order,
-    getSortableRowValues,
-  ]);
+  }, [allResources, filters, orderById, order, orderBy, getSortableRowValues]);
 
   const visibleRows = React.useMemo(
     () =>
@@ -135,14 +177,6 @@ export const ResourcesTable: FC<ResourcesTableProps> = ({
     },
     [order, orderBy],
   );
-
-  const handleSearchChange = useCallback(
-    (event: React.ChangeEvent<HTMLInputElement>) =>
-      setSearchValue(event.target.value),
-    [],
-  );
-
-  const handleSearchClear = useCallback(() => setSearchValue(''), []);
 
   const handleChangePage = useCallback((_event: unknown, newPage: number) => {
     setPage(newPage);
@@ -166,11 +200,10 @@ export const ResourcesTable: FC<ResourcesTableProps> = ({
       id="search-input-filter-toolbar"
       style={{ display: 'flex', alignItems: 'center' }}
     >
-      <ResourcesFilterBy setFilterValue={setFilterValue} />
-      <ResourcesSearchBar
-        value={searchValue}
-        onChange={handleSearchChange}
-        onSearchClear={handleSearchClear}
+      <ResourcesFilterBy
+        filters={filters}
+        setFilters={setFilters}
+        allKinds={getAllKinds}
       />
     </div>
   );
@@ -187,7 +220,7 @@ export const ResourcesTable: FC<ResourcesTableProps> = ({
         />
         {visibleRows.length > 0 ? (
           <TableBody>
-            <ResourcesTableBody rows={visibleRows} createdAt={createdAt} />
+            <ResourcesTableBody rows={visibleRows} />
             {emptyRows > 0 && (
               <TableRow style={{ height: 55 * emptyRows }}>
                 <TableCell colSpan={ResourcesColumnHeaders.length} />