+++ workspaces/redhat-argocd/plugins/argocd/src/components/DeploymentLifeCycle/sidebar/resources/ResourcesTableRow.tsx
@@ -1,4 +1,19 @@
-import React from 'react';
+/*
+ * Copyright 2024 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import React, { useMemo } from 'react';
 import {
   Box,
   Collapse,
@@ -13,48 +28,46 @@ import KeyboardArrowUpIcon from '@mui/icons-material/KeyboardArrowUp';
 
 import { ResourceSyncStatus } from './ResourcesSyncStatus';
 import { ResourceHealthStatus } from './ResourcesHealthStatus';
-import { OpenRowStatus, Resource } from '../../../../types/application';
+import { Resource } from '../../../../types/application';
 import ResourceMetadata from './resource/ResourceMetadata';
 
 type ResourcesTableRowProps = {
-  row: Resource;
-  createdAt: string;
-  open: boolean;
   uid: string;
-  setOpen: React.Dispatch<React.SetStateAction<OpenRowStatus>>;
+  row: Resource;
 };
 
 const useStyles = makeStyles(theme => ({
   resourceRow: {
     '& > *': { borderBottom: 'unset' },
-    '&:first-child': {
-      // borderTop: `1px solid ${theme.palette.grey.A100}`,
-    },
     '&:nth-of-type(odd)': {
       backgroundColor: 'inherit',
     },
-    // boxShadow: '0px 0.5px 2px rgba(0, 0, 0, 0.3)',
-    // borderBottom: `1px solid ${theme.palette.grey.A100}`,
   },
   expandedRow: {
-    '&:first-child': {
-      // borderBottom: `1px solid ${theme.palette.grey.A100}`,
-    },
     padding: 10,
   },
   tableCell: {
-    padding: theme.spacing(1, 2, 1),
+    padding: theme.spacing(1, 0),
     height: 50,
+  },
+  icon: {
+    marginLeft: theme.spacing(0.6),
+    width: '0.8em',
+    height: '0.8em',
   },
 }));
 
 export const ResourcesTableRow: React.FC<ResourcesTableRowProps> = ({
   uid,
   row,
-  createdAt,
 }) => {
   const classes = useStyles();
   const [open, setOpen] = React.useState(false);
+
+  const formattedTimestamp = useMemo(
+    () => moment.utc(row?.createTimestamp).local().format('MM/DD/YYYY hh:mm a'),
+    [row?.createTimestamp],
+  );
 
   return (
     <>
@@ -70,16 +83,21 @@ export const ResourcesTableRow: React.FC<ResourcesTableRowProps> = ({
           </IconButton>
         </TableCell>
         <TableCell className={classes.tableCell} align="left">
+          {row.name}
+        </TableCell>
+        <TableCell className={classes.tableCell} align="left">
           {row.kind}
         </TableCell>
         <TableCell className={classes.tableCell} align="left">
-          {moment.utc(createdAt).local().format('MM/DD/YYYY hh:mm a')}
+          {row?.createTimestamp ? formattedTimestamp : <>-</>}
         </TableCell>
         <TableCell className={classes.tableCell} align="left">
-          <ResourceSyncStatus syncStatus={row.status} />
+          {row.status && <ResourceSyncStatus syncStatus={row.status} />}
         </TableCell>
         <TableCell className={classes.tableCell} align="left">
-          <ResourceHealthStatus healthStatus={row.health?.status} />
+          {row.health?.status && (
+            <ResourceHealthStatus healthStatus={row.health?.status} />
+          )}
         </TableCell>
       </TableRow>
       <TableRow className={classes.expandedRow}>