+++ workspaces/redhat-argocd/plugins/argocd/src/components/DeploymentLifeCycle/sidebar/resources/__tests__/ResourcesTable.test.tsx
@@ -1,8 +1,29 @@
+/*
+ * Copyright 2024 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import React from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 
 import { ResourcesTable } from '../ResourcesTable';
 import { Resource } from '../../../../../types/application';
+import { useArgoResources } from '../../rollouts/RolloutContext';
+
+jest.mock('../../../sidebar/rollouts/RolloutContext', () => ({
+  ...jest.requireActual('../../../sidebar/rollouts/RolloutContext'),
+  useArgoResources: jest.fn(),
+}));
 
 jest.mock('../ResourcesSearchBar', () => ({
   ResourcesSearchBar: jest.fn(({ value, onChange, onSearchClear }) => (
@@ -18,7 +39,7 @@ jest.mock('../ResourcesSearchBar', () => ({
   )),
 }));
 
-jest.mock('../ResourcesFilterBy', () => ({
+jest.mock('../filters/ResourcesFilterBy', () => ({
   ResourcesFilterBy: jest.fn(({ setFilterValue }) => (
     <select
       data-testid="filter-by"
@@ -32,6 +53,11 @@ jest.mock('../ResourcesFilterBy', () => ({
 }));
 
 describe('ResourcesTable Component', () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+    (useArgoResources as jest.Mock).mockReturnValue({ argoResources: [] });
+  });
+
   const resources = [
     {
       kind: 'Rollout',
@@ -104,12 +130,10 @@ describe('ResourcesTable Component', () => {
   const setup = () =>
     render(<ResourcesTable resources={resources} createdAt={createdAt} />);
 
-  it('should render the search bar, filter dropdown, and pagination', () => {
+  it('should render the filter dropdown, and pagination', () => {
     setup();
 
-    expect(screen.getByTestId('search-input')).toBeInTheDocument();
     expect(screen.getByTestId('filter-by')).toBeInTheDocument();
-    expect(screen.getByRole('button', { name: 'Clear' })).toBeInTheDocument();
     expect(screen.getByText('5 rows')).toBeInTheDocument();
   });
 
@@ -118,22 +142,6 @@ describe('ResourcesTable Component', () => {
 
     expect(screen.getByText('Health status')).toBeInTheDocument();
     expect(screen.getByText('Deployment')).toBeInTheDocument();
-  });
-
-  it('should handle search input changes correctly', () => {
-    setup();
-
-    const searchInput = screen.getByTestId('search-input');
-    fireEvent.change(searchInput, { target: { value: 'deployment' } });
-    expect(searchInput).toHaveValue('deployment');
-  });
-
-  it('should handle filter changes correctly', () => {
-    setup();
-
-    const filterDropdown = screen.getByTestId('filter-by');
-    fireEvent.change(filterDropdown, { target: { value: 'Synced' } });
-    expect(filterDropdown).toHaveValue('Synced');
   });
 
   it('should display "No Resources found" when there are no visible rows', () => {
@@ -182,19 +190,6 @@ describe('ResourcesTable Component', () => {
       const newPageResources = screen.getAllByText('Deployment');
       expect(newPageResources.length).toBe(5);
     });
-  });
-
-  it('should handle search clear correctly', () => {
-    setup();
-
-    const searchInput = screen.getByTestId('search-input');
-    const clearButton = screen.getByRole('button', { name: 'Clear' });
-
-    fireEvent.change(searchInput, { target: { value: 'deployment' } });
-    expect(searchInput).toHaveValue('deployment');
-
-    fireEvent.click(clearButton);
-    expect(searchInput).toHaveValue('');
   });
 
   it('should update sort order when handleRequestSort is triggered', async () => {