+++ workspaces/redhat-argocd/plugins/argocd/src/components/DeploymentLifeCycle/sidebar/resources/__tests__/ResourcesTableHeader.test.tsx
@@ -1,4 +1,19 @@
-/* eslint-disable jest/no-conditional-expect */
+/*
+ * Copyright 2024 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 
@@ -49,16 +64,15 @@ describe('ResourcesTableHeader Component', () => {
     const nameHeader = screen.getByText('Name');
     const sortLabel = nameHeader.closest('span');
 
-    if (sortLabel) {
-      fireEvent.click(sortLabel);
-      expect(onRequestSortMock).toHaveBeenCalledWith(
-        expect.any(Object),
-        'Name',
-        '0',
-      );
-    } else {
+    if (!sortLabel) {
       throw new Error('Sort label not found');
     }
+    fireEvent.click(sortLabel);
+    expect(onRequestSortMock).toHaveBeenCalledWith(
+      expect.any(Object),
+      'name',
+      '0',
+    );
   });
 
   it('should not render TableSortLabel for expander column', () => {