+++ workspaces/redhat-argocd/plugins/argocd/src/hooks/useArgoRollouts.ts
@@ -1,16 +1,28 @@
+/*
+ * Copyright 2024 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import { useMemo } from 'react';
-
 import { useEntity } from '@backstage/plugin-catalog-react';
-
 import { useKubernetesObjects } from '@janus-idp/shared-react';
+import pluralize from 'pluralize';
 
 import { kubernetesApiRef, kubernetesAuthProvidersApiRef } from '../kubeApi';
 import {
   ArgoCDkindPluralMap,
   ArgoCDResourcesKind,
   ArgoResources,
-  customResourceKinds,
-  k8sResourceTypes,
 } from '../types/resources';
 
 export const useArgocdRollouts = (): ArgoResources => {
@@ -34,14 +46,15 @@ export const useArgocdRollouts = (): ArgoResources => {
   const argoResources = useMemo<ArgoResources>(() => {
     return (kubernetesObjects?.items?.[0]?.resources || []).reduce(
       (acc, resource) => {
-        if (k8sResourceTypes.includes(resource.type)) {
+        if (resource.type === 'customresources') {
+          const kind = resource.resources?.[0]?.kind as ArgoCDResourcesKind;
+          if (kind) {
+            acc[pluralize(kind).toLocaleLowerCase('en-US')] =
+              resource.resources;
+          }
+        } else {
           const knownK8sType = resource.type as keyof ArgoResources;
           acc[knownK8sType] = resource.resources as any[];
-        } else if (resource.type === 'customresources') {
-          const kind = resource.resources?.[0]?.kind as ArgoCDResourcesKind;
-          if (customResourceKinds.includes(kind)) {
-            acc[ArgoCDkindPluralMap[kind]] = resource.resources;
-          }
         }
         return acc;
       },