+++ workspaces/3scale/plugins/3scale-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## @janus-idp/backstage-plugin-3scale-backend [1.8.0](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-3scale-backend@1.7.1...@janus-idp/backstage-plugin-3scale-backend@1.8.0) (2024-07-25)
 
+## 1.8.5
+
+### Patch Changes
+
+- c6e87b8: Backstage version bump to v1.31.1
+
 ## 1.8.4
 
 ### Patch Changes