+++ workspaces/kafka/plugins/kafka-backend/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage-community/plugin-kafka-backend
 
+## 0.3.20
+
+### Patch Changes
+
+- ec791d1: Backstage version bump to v1.31.1
+
 ## 0.3.19
 
 ### Patch Changes